+++ src/main/java/uk/gov/companieshouse/digitalcertifiedcopyprocessor/service/NullService.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.digitalcertifiedcopyprocessor.consumer;
+package uk.gov.companieshouse.digitalcertifiedcopyprocessor.service;
 
 import org.springframework.stereotype.Component;
 
@@ -6,10 +6,10 @@ import org.springframework.stereotype.Component;
  * The default service is a placeholder for actually hooking up to the consuming service.
  */
 @Component
-class NullService implements Service {
+public class NullService implements KafkaService {
 
     @Override
-    public void processMessage(ServiceParameters parameters) {
+    public void processMessage(KafkaServiceParameters parameters) {
         System.out.println("DATA: "+parameters.getData());
     }
 }
